refactor: add a separate module for consistency helpers
